feat: Mapping entities (User, UserType)

+++ src/main/resources/db/migration/V1__create_table_users.sql
@@ -4,5 +4,5 @@ CREATE TABLE users (
     cpf VARCHAR(11) NOT NULL UNIQUE,
     email VARCHAR(255) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL,
-    balance NUMERIC(15, 2) NOT NULL
+    balance BIGINT NOT NULL
 )
